python-tutorial/requirements.txt
colour==0.1.5
deepface==0.0.87
extcolors==1.0.0
face_recognition==1.3.0
fastapi==0.111.0
healthcheck==1.3.3
inquirer==3.3.0
numpy==1.23.4
opencv_python==4.9.0.80
Pillow==10.4.0
pydantic==2.8.0
pydantic_settings==2.3.4
pymongo==4.6.2
pytesseract==0.3.10
Requests==2.32.3
